requirements.txt
bs4
lxml
Pillow
prettytable
pyyaml
requests
openpyxl